Domain-Specific Vector Embedding Optimization/requirements.txt
chromadb
groq
sentence-transformers
llama-index
python-dotenv
requests
bs4
transformers
llama-index-embeddings-huggingface
llama-index-llms-groq
llama-index-vector-stores-chroma